+++ packages/react-network/src/test/hooks.test.tsx
@@ -4,7 +4,7 @@ import {Header} from '@shopify/network';
 import {FirstArgument} from '@shopify/useful-types';
 import {NetworkManager} from '../manager';
 import {NetworkContext} from '../context';
-import {useCookie, useAcceptLanguage} from '../hooks';
+import {useAcceptLanguage} from '../hooks';
 
 describe('hooks', () => {
   describe('useAcceptLanguage()', () => {
@@ -46,50 +46,6 @@ describe('hooks', () => {
       const wrapper = await mount(<MockComponent />);
 
       expect(wrapper).toContainReactText('en');
-    });
-  });
-
-  describe('useCookie', () => {
-    function MockComponent({cookie}: {cookie: string}) {
-      const [value, setCookie] = useCookie(cookie);
-
-      return (
-        <>
-          <button type="button" onClick={() => setCookie('baz')}>
-            Set Cookie
-          </button>
-          {value}
-        </>
-      );
-    }
-
-    it('gets a cookie', async () => {
-      const key = 'foo';
-      const value = 'bar';
-      const cookies = {[key]: value};
-
-      const wrapper = await mount(<MockComponent cookie={key} />, {
-        manager: new NetworkManager({cookies}),
-      });
-
-      expect(wrapper).toContainReactText(value);
-    });
-
-    it('sets a cookie', async () => {
-      const key = 'foo';
-      const value = 'bar';
-      const cookies = {[key]: value};
-
-      const wrapper = await mount(<MockComponent cookie={key} />, {
-        manager: new NetworkManager({cookies}),
-      });
-
-      wrapper
-        .find(MockComponent)!
-        .find('button')!
-        .trigger('onClick');
-
-      expect(wrapper).toContainReactText(`baz`);
     });
   });
 });
